Introduces an `absolute_path` method
Problem:
There were problems with Sidekiq workers accessing the file system and
trying to access directories which could not be found.

Solution:
It is uncertain if this change fixes the problem. Nevertheless a new
method is introduced to make sure the worker uses a correct (absolute)
path to the desired file/dir.

QB-494